Moved contigInsertSize
